pyproject.toml

[tool.setuptools.packages.find]
where = ["src"]

[tool.poetry]
name = "linear-cae"
version = "0.1.4"
description = "Linear Consistency Autoencoders"
authors = [
    "Bernardo Torres",
    "Gabriel Meseguer-Brocal",
]
packages = [{include = "linear_cae", from = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
numpy = "^1.23.5"
matplotlib = ">=3.10.1"
torch = ">=2.0.1"
soundfile = ">=0.13.1"
huggingface_hub = "*"
PyYAML = "*"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.baselines]
optional = true

[tool.poetry.group.stable-audio]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
ruff = "^0.14.0"
torch = ">=2.7.0"
torchaudio = ">=2.0.1"
lightning = ">=2.5.1.post0,<3.0.0"
ipython = "^9.3.0"
musdb = "^0.4.3"
ipykernel = "*"
wandb = ">=0.19.9,<0.20.0"
auraloss = ">=0.4.0,<0.5.0"
hydra-core = ">=1.3.2,<2.0.0"
holoviews = ">=1.20.2,<2.0.0"
tzlocal = ">=5.3.1,<6.0.0"
seaborn = ">=0.13.2,<0.14.0"
rich = "^14.0.0"
rootutils = ">=1.0.7,<2.0.0"
hydra-colorlog = ">=1.2.0,<2.0.0"
bokeh = ">=3.7.2,<4.0.0"
gputil = "^1.4.0"
resampy = ">=0.4.3,<0.5.0"
librosa = ">=0.11.0"
scipy = ">=1.15.2,<2.0.0"
wavetable = ">=0.1.5"

[tool.poetry.group.baselines.dependencies]
tensorboard = ">=2.19.0,<3.0.0"
music2latent = "*"
transformers = "^4.56.0"
diffusers = "0.25.0"

# [tool.poetry.group.stable-audio.dependencies]
# stable-audio-tools = "^0.0.19"


[tool.poetry.scripts]
m2l = "linear_cae.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.black]
line-length = 88
target-version = ['py310']


[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort (import sorting)
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"] # Ignore line too long, handled by black

[tool.ruff.isort]
known-first-party = ["linear_cae"]

[tool.pytest.ini_options]
addopts = "--cov-report xml:coverage.xml --cov=. --junitxml=xunit.xml"
junit_family = "xunit2"
